% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{BoundedHazard_ExponentialDistribution}
\alias{BoundedHazard_ExponentialDistribution}
\title{ExponentialDistribution: BoundedHazard}
\usage{
BoundedHazard_ExponentialDistribution()
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ExponentialDistribution$BoundedHazard}
}
}
