% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Gamma.R
\name{BoundedHazard_GammaDistribution}
\alias{BoundedHazard_GammaDistribution}
\title{GammaDistribution: Bounded Hazard}
\usage{
BoundedHazard_GammaDistribution()
}
\description{
im a method!
\itemize{
\item This method is bound to \code{GammaDistribution$BoundedHazard}
}
}
