% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{BoundedHazard_ShiftedExponentialDistribution}
\alias{BoundedHazard_ShiftedExponentialDistribution}
\title{ShiftedExponentialDistribution: Bounded Hazard}
\usage{
BoundedHazard_ShiftedExponentialDistribution()
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ShiftedExponentialDistribution$BoundedHazard}
}
}
