% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Weibull.R
\name{BoundedHazard_WeibullDistribution}
\alias{BoundedHazard_WeibullDistribution}
\title{WeibullDistribution: Bounded Hazard}
\usage{
BoundedHazard_WeibullDistribution()
}
\description{
im a method!
\itemize{
\item This method is bound to \code{WeibullDistribution$BoundedHazard}
}
}
