% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{CheckFracError}
\alias{CheckFracError}
\title{Check Fractional Error}
\usage{
CheckFracError(a, b, tol, m)
}
\arguments{
\item{a}{numeric}

\item{b}{numeric}

\item{tol}{numeric}

\item{m}{character}
}
\description{
If the fractional error \deqn{\frac{\left | a-b \right |}{b} > tol} return \code{FALSE} and alert the user, otherwise return \code{TRUE}
\itemize{
\item This method is bound to \code{Distribution$CheckFracError} for
\itemize{
\item \code{\link{ExponentialDistribution}}
\item \code{\link{ShiftedExponentialDistribution}}
\item \code{\link{WeibullDistribution}}
}
}
}
