% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Gamma.R
\name{CheckSamples_GammaDistribution}
\alias{CheckSamples_GammaDistribution}
\title{GammaDistribution: Check Samples}
\usage{
CheckSamples_GammaDistribution(samples, dt)
}
\arguments{
\item{samples}{numeric vector}

\item{dt}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{GammaDistribution$CheckSamples}
}
}
