% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{CheckSamples_ShiftedExponentialDistribution}
\alias{CheckSamples_ShiftedExponentialDistribution}
\title{ShiftedExponentialDistribution: Check Samples}
\usage{
CheckSamples_ShiftedExponentialDistribution(samples, dt)
}
\arguments{
\item{samples}{numeric vector}

\item{dt}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ShiftedExponentialDistribution$CheckSamples}
}
}
