% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Weibull.R
\name{CheckSamples_WeibullDistribution}
\alias{CheckSamples_WeibullDistribution}
\title{WeibullDistribution: Check Samples}
\usage{
CheckSamples_WeibullDistribution(samples, dt)
}
\arguments{
\item{samples}{numeric vector}

\item{dt}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{WeibullDistribution$CheckSamples}
}
}
