% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Dirac.R
\docType{class}
\name{DiracDistribution}
\alias{DiracDistribution}
\title{DiracDistribution Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
DiracDistribution
}
\description{
do something
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item method: im a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item field: im a field!
}
}

\keyword{R6}
\keyword{class}
