% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{FracError}
\alias{FracError}
\title{Fractional Error}
\usage{
FracError(a, b)
}
\arguments{
\item{a}{numeric}

\item{b}{numeric}
}
\description{
Used to in \code{\link{CheckFracError}}
\itemize{
\item This method is bound to \code{Distribution$FracError} for:
\itemize{
\item \code{\link{ExponentialDistribution}}
\item \code{\link{ShiftedExponentialDistribution}}
}
}
}
