% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gamma_p_boost}
\alias{gamma_p_boost}
\title{Normalised Lower Incomplete Gamma Function}
\usage{
gamma_p_boost(a, z)
}
\arguments{
\item{a}{numeric}

\item{z}{numeric}
}
\description{
Return the normalised lower incomplete gamma function; this is a wrapper for
\code{gamma_p}, found in \url{http://www.boost.org/doc/libs/1_46_0/libs/math/doc/sf_and_dist/html/math_toolkit/special/sf_gamma/igamma.html}
}
