% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Gamma.R
\name{get_alpha_GammaDistribution}
\alias{get_alpha_GammaDistribution}
\title{GammaDistribution: Get \code{alpha}}
\usage{
get_alpha_GammaDistribution()
}
\description{
im a method!
\itemize{
\item This method is bound to \code{GammaDistribution$get_alpha}
}
}
