% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Gamma.R
\name{get_beta_GammaDistribution}
\alias{get_beta_GammaDistribution}
\title{GammaDistribution: Get \code{beta}}
\usage{
get_beta_GammaDistribution()
}
\description{
im a method!
\itemize{
\item This method is bound to \code{GammaDistribution$get_beta}
}
}
