% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Dirac.R
\name{get_enabling_time_DiracDistribution}
\alias{get_enabling_time_DiracDistribution}
\title{DiracDistribution: Get \code{enabling_time}}
\usage{
get_enabling_time_DiracDistribution()
}
\description{
im a method!
\itemize{
\item This method is bound to \code{DiracDistribution$get_enabling_time}
}
}
