% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{get_enabling_time_ExponentialDistribution}
\alias{get_enabling_time_ExponentialDistribution}
\title{ExponentialDistribution: Get \code{enabling_time}}
\usage{
get_enabling_time_ExponentialDistribution()
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ExponentialDistribution$get_enabling_time}
}
}
