% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Gamma.R
\name{get_enabling_time_GammaDistribution}
\alias{get_enabling_time_GammaDistribution}
\title{GammaDistribution: Get \code{enabling_time}}
\usage{
get_enabling_time_GammaDistribution()
}
\description{
im a method!
\itemize{
\item This method is bound to \code{GammaDistribution$get_enabling_time}
}
}
