% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{get_enabling_time_ShiftedExponentialDistribution}
\alias{get_enabling_time_ShiftedExponentialDistribution}
\title{ShiftedExponentialDistribution: Get \code{enabling_time}}
\usage{
get_enabling_time_ShiftedExponentialDistribution()
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ShiftedExponentialDistribution$get_enabling_time}
}
}
