% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Weibull.R
\name{get_k_WeibullDistribution}
\alias{get_k_WeibullDistribution}
\title{WeibullDistribution: Get \code{k}}
\usage{
get_k_WeibullDistribution()
}
\description{
im a method!
\itemize{
\item This method is bound to \code{WeibullDistribution$get_k}
}
}
