% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{get_lambda_ExponentialDistribution}
\alias{get_lambda_ExponentialDistribution}
\title{ExponentialDistribution: Get \code{lambda}}
\usage{
get_lambda_ExponentialDistribution()
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ExponentialDistribution$get_lambda}
}
}
