% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{get_lambda_ShiftedExponentialDistribution}
\alias{get_lambda_ShiftedExponentialDistribution}
\title{ShiftedExponentialDistribution: Get \code{lambda}}
\usage{
get_lambda_ShiftedExponentialDistribution()
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ShiftedExponentialDistribution$get_lambda}
}
}
