% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Weibull.R
\name{get_lambda_WeibullDistribution}
\alias{get_lambda_WeibullDistribution}
\title{WeibullDistribution: Get \code{lambda}}
\usage{
get_lambda_WeibullDistribution()
}
\description{
im a method!
\itemize{
\item This method is bound to \code{WeibullDistribution$get_lambda}
}
}
