% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{get_normal_ExponentialDistribution}
\alias{get_normal_ExponentialDistribution}
\title{ExponentialDistribution: Get \code{normal}}
\usage{
get_normal_ExponentialDistribution()
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ExponentialDistribution$get_normal}
}
}
