% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{get_normal_ShiftedExponentialDistribution}
\alias{get_normal_ShiftedExponentialDistribution}
\title{ShiftedExponentialDistribution: Get \code{normal}}
\usage{
get_normal_ShiftedExponentialDistribution()
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ShiftedExponentialDistribution$get_normal}
}
}
