% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Weibull.R
\name{get_normal_WeibullDistribution}
\alias{get_normal_WeibullDistribution}
\title{WeibullDistribution: Get \code{normal}}
\usage{
get_normal_WeibullDistribution()
}
\description{
im a method!
\itemize{
\item This method is bound to \code{WeibullDistribution$get_normal}
}
}
