% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{get_shift_ShiftedExponentialDistribution}
\alias{get_shift_ShiftedExponentialDistribution}
\title{ShiftedExponentialDistribution: Get \code{shift}}
\usage{
get_shift_ShiftedExponentialDistribution()
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ShiftedExponentialDistribution$get_shift}
}
}
