% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Dirac.R
\name{get_value_DiracDistribution}
\alias{get_value_DiracDistribution}
\title{DiracDistribution: Get \code{alpha}}
\usage{
get_value_DiracDistribution()
}
\description{
im a method!
\itemize{
\item This method is bound to \code{DiracDistribution$get_value}
}
}
