% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{HazardIntegral_ExponentialDistribution}
\alias{HazardIntegral_ExponentialDistribution}
\title{ExponentialDistribution: HazardIntegral}
\usage{
HazardIntegral_ExponentialDistribution(current_time)
}
\arguments{
\item{t0}{numeric}

\item{t1}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ExponentialDistribution$HazardIntegral}
}
}
