% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Gamma.R
\name{HazardIntegral_GammaDistribution}
\alias{HazardIntegral_GammaDistribution}
\title{GammaDistribution: Hazard Integral}
\usage{
HazardIntegral_GammaDistribution(t0, t1)
}
\arguments{
\item{t0}{numeric}

\item{t1}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{GammaDistribution$HazardIntegral}
}
}
