% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Weibull.R
\name{HazardIntegral_WeibullDistribution}
\alias{HazardIntegral_WeibullDistribution}
\title{WeibullDistribution: Hazard Integral}
\usage{
HazardIntegral_WeibullDistribution(t0, t1)
}
\arguments{
\item{t0}{numeric}

\item{t1}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{WeibullDistribution$HazardIntegral}
}
}
