% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{ImplicitHazardIntegral_ExponentialDistribution}
\alias{ImplicitHazardIntegral_ExponentialDistribution}
\title{ExponentialDistribution: Implicit Hazard Integral}
\usage{
ImplicitHazardIntegral_ExponentialDistribution(current_time)
}
\arguments{
\item{xa}{numeric}

\item{t0}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ExponentialDistribution$ImplicitHazardIntegral}
}
}
