% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Gamma.R
\name{ImplicitHazardIntegral_GammaDistribution}
\alias{ImplicitHazardIntegral_GammaDistribution}
\title{GammaDistribution: Implicit Hazard Integral}
\usage{
ImplicitHazardIntegral_GammaDistribution(xa, t0)
}
\arguments{
\item{xa}{numeric}

\item{t0}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{GammaDistribution$ImplicitHazardIntegral}
}
}
