% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{ImplicitHazardIntegral_ShiftedExponentialDistribution}
\alias{ImplicitHazardIntegral_ShiftedExponentialDistribution}
\title{ShiftedExponentialDistribution: Implicit Hazard Integral}
\usage{
ImplicitHazardIntegral_ShiftedExponentialDistribution(xa, t0)
}
\arguments{
\item{xa}{numeric}

\item{t0}{numeric}
}
\description{
This implicitly solves for a quantile by integrating the hazard. The function returns \eqn{t} in \deqn{x_{a}=\int_{t0}^{t}\lambda (s)ds}
\itemize{
\item This method is bound to \code{ShiftedExponentialDistribution$ImplicitHazardIntegral}
}
}
