% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Weibull.R
\name{ImplicitHazardIntegral_WeibullDistribution}
\alias{ImplicitHazardIntegral_WeibullDistribution}
\title{WeibullDistribution: ImplicitHazardIntegral}
\usage{
ImplicitHazardIntegral_WeibullDistribution(xa, t0)
}
\arguments{
\item{xa}{numeric}

\item{t0}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{WeibullDistribution$ImplicitHazardIntegral}
}
}
