\name{print.hazard}
\alias{print.hazard}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print method for object of class \code{hazard}
}
\description{
Print method for objects of class \code{hazard}.
}
\usage{
\method{print}{hazard}(x, whole = FALSE, ...)
}
\arguments{
  \item{x}{
  An object of class \code{hazard}.
}
  \item{whole}{
  A logical value indicating if the whole vectors of the object \code{hazard} should be displayed. Default is \code{FALSE} which means that only the first six values are returned.
  }
  \item{\dots}{
Further arguments for print or summary.
}
}

%\details{
%The number of objects displayed depends on how many arguments were used in \link{hazard}.
%}

\value{
No value returned.
}

\author{
Agnieszka Listwon-Krol
}

\seealso{
\link{hazard}, \link{plot.hazard}
}

\keyword{documentation}
