% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Dirac.R
\name{Sample_DiracDistribution}
\alias{Sample_DiracDistribution}
\title{DiracDistribution: Sample}
\usage{
Sample_DiracDistribution(current_time)
}
\arguments{
\item{current_time}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{DiracDistribution$Sample}
}
}
