% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{Sample_ExponentialDistribution}
\alias{Sample_ExponentialDistribution}
\title{ExponentialDistribution: Sample}
\usage{
Sample_ExponentialDistribution(current_time)
}
\arguments{
\item{current_time}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ExponentialDistribution$Sample}
}
}
