% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Gamma.R
\name{Sample_GammaDistribution}
\alias{Sample_GammaDistribution}
\title{GammaDistribution: Sample}
\usage{
Sample_GammaDistribution(current_time)
}
\arguments{
\item{current_time}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{GammaDistribution$Sample}
}
}
