% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{Sample_ShiftedExponentialDistribution}
\alias{Sample_ShiftedExponentialDistribution}
\title{ShiftedExponentialDistribution: Sample}
\usage{
Sample_ShiftedExponentialDistribution(current_time)
}
\arguments{
\item{current_time}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ShiftedExponentialDistribution$Sample}
}
}
