% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Weibull.R
\name{Sample_WeibullDistribution}
\alias{Sample_WeibullDistribution}
\title{WeibullDistribution: Sample}
\usage{
Sample_WeibullDistribution(current_time)
}
\arguments{
\item{current_time}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{WeibullDistribution$Sample}
}
}
