% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Dirac.R
\name{set_enabling_time_DiracDistribution}
\alias{set_enabling_time_DiracDistribution}
\title{DiracDistribution: Set \code{enabling_time}}
\usage{
set_enabling_time_DiracDistribution(enabling_time)
}
\arguments{
\item{enabling_time}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{DiracDistributio$get_enabling_time}
}
}
