% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Gamma.R
\name{set_enabling_time_GammaDistribution}
\alias{set_enabling_time_GammaDistribution}
\title{GammaDistribution: Set \code{enabling_time}}
\usage{
set_enabling_time_GammaDistribution(enabling_time)
}
\arguments{
\item{enabling_time}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{GammaDistribution$get_enabling_time}
}
}
