% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{set_enabling_time_ShiftedExponentialDistribution}
\alias{set_enabling_time_ShiftedExponentialDistribution}
\title{ShiftedExponentialDistribution: Set \code{enabling_time}}
\usage{
set_enabling_time_ShiftedExponentialDistribution(enabling_time)
}
\arguments{
\item{enabling_time}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ShiftedExponentialDistribution$get_enabling_time}
}
}
