% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Weibull.R
\name{set_enabling_time_WeibullDistribution}
\alias{set_enabling_time_WeibullDistribution}
\title{WeibullDistribution: Set \code{enabling_time}}
\usage{
set_enabling_time_WeibullDistribution(enabling_time)
}
\arguments{
\item{enabling_time}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{WeibullDistribution$get_enabling_time}
}
}
