% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Weibull.R
\name{set_k_WeibullDistribution}
\alias{set_k_WeibullDistribution}
\title{WeibullDistribution: Set \code{k}}
\usage{
set_k_WeibullDistribution(k)
}
\arguments{
\item{k}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{WeibullDistribution$get_k}
}
}
