% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{set_lambda_ExponentialDistribution}
\alias{set_lambda_ExponentialDistribution}
\title{ExponentialDistribution: Set \code{lambda}}
\usage{
set_lambda_ExponentialDistribution(lambda)
}
\arguments{
\item{lambda}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ExponentialDistribution$set_lambda}
}
}
