% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{set_lambda_ShiftedExponentialDistribution}
\alias{set_lambda_ShiftedExponentialDistribution}
\title{ShiftedExponentialDistribution: Set \code{lambda}}
\usage{
set_lambda_ShiftedExponentialDistribution(lambda)
}
\arguments{
\item{lambda}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ShiftedExponentialDistribution$set_lambda}
}
}
