% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Weibull.R
\name{set_lambda_WeibullDistribution}
\alias{set_lambda_WeibullDistribution}
\title{WeibullDistribution: Set \code{lambda}}
\usage{
set_lambda_WeibullDistribution(lambda)
}
\arguments{
\item{lambda}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{WeibullDistribution$get_lambda}
}
}
