% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{set_normal_ExponentialDistribution}
\alias{set_normal_ExponentialDistribution}
\title{ExponentialDistribution: Set \code{normal}}
\usage{
set_normal_ExponentialDistribution(normal)
}
\arguments{
\item{lambda}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ExponentialDistribution$set_normal}
}
}
