% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{set_normal_ShiftedExponentialDistribution}
\alias{set_normal_ShiftedExponentialDistribution}
\title{ShiftedExponentialDistribution: Set \code{normal}}
\usage{
set_normal_ShiftedExponentialDistribution(normal)
}
\arguments{
\item{lambda}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ShiftedExponentialDistribution$set_normal}
}
}
