% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Weibull.R
\name{set_normal_WeibullDistribution}
\alias{set_normal_WeibullDistribution}
\title{WeibullDistribution: Set \code{normal}}
\usage{
set_normal_WeibullDistribution(normal)
}
\arguments{
\item{normal}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{WeibullDistribution$get_normal}
}
}
