% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{set_shift_ShiftedExponentialDistribution}
\alias{set_shift_ShiftedExponentialDistribution}
\title{ShiftedExponentialDistribution: Set \code{shift}}
\usage{
set_shift_ShiftedExponentialDistribution(shift)
}
\arguments{
\item{lambda}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ShiftedExponentialDistribution$set_shift}
}
}
