% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Weibull.R
\name{set_shift_WeibullDistribution}
\alias{set_shift_WeibullDistribution}
\title{WeibullDistribution: Set \code{shift}}
\usage{
set_shift_WeibullDistribution(shift)
}
\arguments{
\item{shift}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{WeibullDistribution$get_shift}
}
}
