% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Dirac.R
\name{set_value_DiracDistribution}
\alias{set_value_DiracDistribution}
\title{DiracDistribution: Set \code{alpha}}
\usage{
set_value_DiracDistribution(value)
}
\arguments{
\item{value}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{DiracDistribution$get_alpha}
}
}
