% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tgamma_boost}
\alias{tgamma_boost}
\title{Full (non-normalised) Upper Incomplete Gamma Function}
\usage{
tgamma_boost(a, z)
}
\arguments{
\item{a}{numeric}

\item{z}{numeric}
}
\description{
Return the full (non-normalised) upper incomplete gamma function; this is a wrapper for
\code{tgamma}, found in \url{http://www.boost.org/doc/libs/1_46_0/libs/math/doc/sf_and_dist/html/math_toolkit/special/sf_gamma/igamma.html}
}
