% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Exponential.R
\name{CheckSamples_ExponentialDistribution}
\alias{CheckSamples_ExponentialDistribution}
\title{ExponentialDistribution: CheckSamples}
\usage{
CheckSamples_ExponentialDistribution(samples, dt)
}
\arguments{
\item{samples}{numeric vector}

\item{dt}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ExponentialDistribution$CheckSamples}
}
}
