% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Weibull.R
\name{EnablingTime_ExponentialDistribution}
\alias{EnablingTime_ExponentialDistribution}
\title{ExponentialDistribution: Enabling Time}
\usage{
EnablingTime_ExponentialDistribution(current_time)
}
\description{
im a method!
\itemize{
\item This method is bound to \code{ExponentialDistribution$EnablingTime}
}
}
