% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Gamma.R
\name{set_alpha_GammaDistribution}
\alias{set_alpha_GammaDistribution}
\title{WeibullDistribution: Set \code{alpha}}
\usage{
set_alpha_GammaDistribution(alpha)
}
\arguments{
\item{alpha}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{WeibullDistribution$get_alpha}
}
}
