% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions-Gamma.R
\name{set_beta_GammaDistribution}
\alias{set_beta_GammaDistribution}
\title{WeibullDistribution: Set \code{beta}}
\usage{
set_beta_GammaDistribution(beta)
}
\arguments{
\item{beta}{numeric}
}
\description{
im a method!
\itemize{
\item This method is bound to \code{WeibullDistribution$get_beta}
}
}
